requirements.txt
imagecodecs==2022.2.22
ipywidgets==7.7.0
jupyterlab==3.4.2
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
plotly==5.8.0
tifffile==2022.5.4
torch==1.11.0 
torchelastic==0.2.2
torchtext==0.12.0
torchvision==0.12.0
hydra-core==1.1.2
wandb==0.12.17
numpy-quaternion>=2021.11.4.15.26.3
click>=8.0
pillow
scipy==1.7.1
imgui==1.3.0
glfw==2.2.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
ninja==1.10.2
prettytable==3.5.0
git+https://github.com/openai/CLIP.git@40f5484c1c74edd83cb9cf687c6ab92b28d8b656
pyspng
psutil
mrcfile
tensorboard
moviepy
kornia 
